+++ codegen/src/persist_table/generator/mod.rs
@@ -8,4 +8,5 @@ mod space_serialize;
 pub struct Generator {
     pub struct_def: ItemStruct,
     pub pk_ident: Ident,
+    pub index_type_ident: Ident
 }