+++ codegen/src/persist_table/mod.rs
@@ -10,10 +10,12 @@ mod parser;
 pub fn expand(input: TokenStream) -> syn::Result<TokenStream> {
     let input_fn = Parser::parse_struct(input)?;
     let pk_ident = Parser::parse_pk_ident(&input_fn);
+    let index_type_ident = Parser::parse_index_ident(&input_fn);
 
     let gen = Generator {
         struct_def: input_fn,
         pk_ident,
+        index_type_ident
     };
 
     let space_type = gen.gen_space_type()?;